fixup! refactor(types): fix type errors
